🐛 fix bug with refactored code
